Safari-609.1.11.1/LayoutTests/fast/dom/HTMLTemplateElement/innerHTML-inert-expected.txt

Blocked access to external URL http://active.com/bar.jpg
Test that setting HTMLTemplateElement.innerHTML uses the inert document when creating elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attemptedLoad is false
PASS successfullyParsed is true

TEST COMPLETE
